Function pallet_identity::pallet::dispatchables::kill_identity

pub fn kill_identity<T: Config>(
    target: <<T as Config>::Lookup as StaticLookup>::Source
)

Remove an account’s identity and sub-account information and slash the deposits.

Payment: Reserved balances from set_subs and set_identity are slashed and handled by Slash. Verification request deposits are not returned; they should be cancelled manually using cancel_request.

The dispatch origin for this call must match T::ForceOrigin.

target: the account whose identity the judgement is upon. This must be an account with a registered identity.

Emits IdentityKilled if successful.

Complexity

O(R + S + X)

where R registrar-count (governance-bounded).

where S subs-count (hard- and deposit-bounded).

where X additional-field-count (deposit-bounded and code-bounded).

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the kill_identity variant of the pallet Call enum. You should not attempt to call this function directly.