Function pallet_identity::pallet::dispatchables::request_judgement

pub fn request_judgement<T: Config>(
    reg_index: RegistrarIndex,
    max_fee: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance
)

Request a judgement from a registrar.

Payment: At most max_fee will be reserved for payment to the registrar if judgement given.

The dispatch origin for this call must be Signed and the sender must have a registered identity.

reg_index: The index of the registrar whose judgement is requested.

max_fee: The maximum fee that may be paid. This should just be auto-populated as:

Self::registrars().get(reg_index).unwrap().fee

Emits JudgementRequested if successful.

Complexity

O(R + X).

where R registrar-count (governance-bounded).

where X additional-field-count (deposit-bounded and code-bounded).

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::request_judgement.