Function pallet_identity::pallet::dispatchables::set_subs

pub fn set_subs<T: Config>(subs: Vec<(T::AccountId, Data)>)

Set the sub-accounts of the sender.

Payment: Any aggregate balance reserved by previous set_subs calls will be returned and an amount SubAccountDeposit will be reserved for each item in subs.

The dispatch origin for this call must be Signed and the sender must have a registered identity.

subs: The identity’s (new) sub-accounts.

Complexity

O(P + S)

where P old-subs-count (hard- and deposit-bounded).

where S subs-count (hard- and deposit-bounded).

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the set_subs variant of the pallet Call enum. You should not attempt to call this function directly.