Enum pallet_identity::pallet::Call

pub enum Call<T: Config> {

    add_registrar {
        account: <<T as Config>::Lookup as StaticLookup>::Source,
    },
    set_identity {
        info: Box<IdentityInfo<T::MaxAdditionalFields>>,
    },
    set_subs {
        subs: Vec<(T::AccountId, Data)>,
    },
    clear_identity {},
    request_judgement {
        reg_index: RegistrarIndex,
        max_fee: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance,
    },
    cancel_request {
        reg_index: RegistrarIndex,
    },
    set_fee {
        index: RegistrarIndex,
        fee: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance,
    },
    set_account_id {
        index: RegistrarIndex,
        new: <<T as Config>::Lookup as StaticLookup>::Source,
    },
    set_fields {
        index: RegistrarIndex,
        fields: IdentityFields,
    },
    provide_judgement {
        reg_index: RegistrarIndex,
        target: <<T as Config>::Lookup as StaticLookup>::Source,
        judgement: Judgement<<<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>,
        identity: T::Hash,
    },
    kill_identity {
        target: <<T as Config>::Lookup as StaticLookup>::Source,
    },
    add_sub {
        sub: <<T as Config>::Lookup as StaticLookup>::Source,
        data: Data,
    },
    rename_sub {
        sub: <<T as Config>::Lookup as StaticLookup>::Source,
        data: Data,
    },
    remove_sub {
        sub: <<T as Config>::Lookup as StaticLookup>::Source,
    },
    quit_sub {},
    // some variants omitted

}

Identity pallet declaration.

Variants

add_registrar

Fields

account: <<T as Config>::Lookup as StaticLookup>::Source

Add a registrar to the system.

The dispatch origin for this call must be T::RegistrarOrigin.

account: the account of the registrar.

Emits RegistrarAdded if successful.

O(R) where R registrar-count (governance-bounded and code-bounded).

One storage mutation (codec O(R)).

One event.

set_identity

Fields

info: Box<IdentityInfo<T::MaxAdditionalFields>>

Set an account’s identity information and reserve the appropriate deposit.

If the account already has identity information, the deposit is taken as part payment for the new deposit.

The dispatch origin for this call must be Signed.

info: The identity information.

Emits IdentitySet if successful.

O(X + X' + R)

where X additional-field-count (deposit-bounded and code-bounded)

where R judgements-count (registrar-count-bounded)

One balance reserve operation.

One storage mutation (codec-read O(X' + R), codec-write O(X + R)).

One event.

set_subs

Fields

subs: Vec<(T::AccountId, Data)>

Set the sub-accounts of the sender.

Payment: Any aggregate balance reserved by previous set_subs calls will be returned and an amount SubAccountDeposit will be reserved for each item in subs.

The dispatch origin for this call must be Signed and the sender must have a registered identity.

subs: The identity’s (new) sub-accounts.

O(P + S)

where P old-subs-count (hard- and deposit-bounded).

where S subs-count (hard- and deposit-bounded).

At most one balance operations.

DB:

P + S storage mutations (codec complexity O(1))

One storage read (codec complexity O(P)).

One storage write (codec complexity O(S)).

One storage-exists (IdentityOf::contains_key).

clear_identity

Fields

Clear an account’s identity info and all sub-accounts and return all deposits.

Payment: All reserved balances on the account are returned.

The dispatch origin for this call must be Signed and the sender must have a registered identity.

Emits IdentityCleared if successful.

O(R + S + X)

where R registrar-count (governance-bounded).

where S subs-count (hard- and deposit-bounded).

where X additional-field-count (deposit-bounded and code-bounded).

One balance-unreserve operation.

2 storage reads and S + 2 storage deletions.

One event.

request_judgement

Fields

reg_index: RegistrarIndex

max_fee: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance

Request a judgement from a registrar.

Payment: At most max_fee will be reserved for payment to the registrar if judgement given.

The dispatch origin for this call must be Signed and the sender must have a registered identity.

reg_index: The index of the registrar whose judgement is requested.

max_fee: The maximum fee that may be paid. This should just be auto-populated as:

Self::registrars().get(reg_index).unwrap().fee

Emits JudgementRequested if successful.

O(R + X).

One balance-reserve operation.

Storage: 1 read O(R), 1 mutate O(X + R).

One event.

cancel_request

Fields

reg_index: RegistrarIndex

Cancel a previous request.

Payment: A previously reserved deposit is returned on success.

The dispatch origin for this call must be Signed and the sender must have a registered identity.

reg_index: The index of the registrar whose judgement is no longer requested.

Emits JudgementUnrequested if successful.

O(R + X).

One balance-reserve operation.

One storage mutation O(R + X).

One event

set_fee

Fields

index: RegistrarIndex

fee: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance

Set the fee required for a judgement to be requested from a registrar.

The dispatch origin for this call must be Signed and the sender must be the account of the registrar whose index is index.

index: the index of the registrar whose fee is to be set.

fee: the new fee.

O(R).

One storage mutation O(R).

Benchmark: 7.315 + R * 0.329 µs (min squares analysis)

set_account_id

Fields

index: RegistrarIndex

new: <<T as Config>::Lookup as StaticLookup>::Source

Change the account associated with a registrar.

The dispatch origin for this call must be Signed and the sender must be the account of the registrar whose index is index.

index: the index of the registrar whose fee is to be set.

new: the new account ID.

O(R).

One storage mutation O(R).

Benchmark: 8.823 + R * 0.32 µs (min squares analysis)

set_fields

Fields

index: RegistrarIndex

fields: IdentityFields

Set the field information for a registrar.

The dispatch origin for this call must be Signed and the sender must be the account of the registrar whose index is index.

index: the index of the registrar whose fee is to be set.

fields: the fields that the registrar concerns themselves with.

O(R).

One storage mutation O(R).

Benchmark: 7.464 + R * 0.325 µs (min squares analysis)

provide_judgement

Fields

reg_index: RegistrarIndex

target: <<T as Config>::Lookup as StaticLookup>::Source

judgement: Judgement<<<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>

identity: T::Hash

Provide a judgement for an account’s identity.

The dispatch origin for this call must be Signed and the sender must be the account of the registrar whose index is reg_index.

reg_index: the index of the registrar whose judgement is being made.

target: the account whose identity the judgement is upon. This must be an account with a registered identity.

judgement: the judgement of the registrar of index reg_index about target.

identity: The hash of the IdentityInfo for that the judgement is provided.

Emits JudgementGiven if successful.

O(R + X).

One balance-transfer operation.

Up to one account-lookup operation.

Storage: 1 read O(R), 1 mutate O(R + X).

One event.

kill_identity

Fields

target: <<T as Config>::Lookup as StaticLookup>::Source

Remove an account’s identity and sub-account information and slash the deposits.

Payment: Reserved balances from set_subs and set_identity are slashed and handled by Slash. Verification request deposits are not returned; they should be cancelled manually using cancel_request.

The dispatch origin for this call must match T::ForceOrigin.

target: the account whose identity the judgement is upon. This must be an account with a registered identity.

Emits IdentityKilled if successful.

O(R + S + X).

One balance-reserve operation.

S + 2 storage mutations.

One event.

add_sub

Fields

sub: <<T as Config>::Lookup as StaticLookup>::Source

data: Data

Add the given account to the sender’s subs.

Payment: Balance reserved by a previous set_subs call for one sub will be repatriated to the sender.

The dispatch origin for this call must be Signed and the sender must have a registered sub identity of sub.

rename_sub

Fields

sub: <<T as Config>::Lookup as StaticLookup>::Source

data: Data

Alter the associated name of the given sub-account.

The dispatch origin for this call must be Signed and the sender must have a registered sub identity of sub.

remove_sub

Fields

sub: <<T as Config>::Lookup as StaticLookup>::Source

Remove the given account from the sender’s subs.

Payment: Balance reserved by a previous set_subs call for one sub will be repatriated to the sender.

The dispatch origin for this call must be Signed and the sender must have a registered sub identity of sub.

quit_sub

Fields

Remove the sender as a sub-account.

Payment: Balance reserved by a previous set_subs call for one sub will be repatriated to the sender (not the original depositor).

The dispatch origin for this call must be Signed and the sender must have a registered super-identity.

NOTE: This should not normally be used, but is provided in the case that the non- controller of an account is maliciously registered as a sub-account.

Implementations

impl<T: Config> Call<T>

pub fn new_call_variant_add_registrar(
account: <<T as Config>::Lookup as StaticLookup>::Source
) -> Self

Create a call with the variant add_registrar.

pub fn new_call_variant_set_identity(
info: Box<IdentityInfo<T::MaxAdditionalFields>>
) -> Self

Create a call with the variant set_identity.

pub fn new_call_variant_set_subs(subs: Vec<(T::AccountId, Data)>) -> Self

Create a call with the variant set_subs.

pub fn new_call_variant_clear_identity() -> Self

Create a call with the variant clear_identity.

pub fn new_call_variant_request_judgement(
reg_index: RegistrarIndex,
max_fee: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance
) -> Self

Create a call with the variant request_judgement.

pub fn new_call_variant_cancel_request(reg_index: RegistrarIndex) -> Self

Create a call with the variant cancel_request.

pub fn new_call_variant_set_fee(
index: RegistrarIndex,
fee: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance
) -> Self

Create a call with the variant set_fee.

pub fn new_call_variant_set_account_id(
index: RegistrarIndex,
new: <<T as Config>::Lookup as StaticLookup>::Source
) -> Self

Create a call with the variant set_account_id.

pub fn new_call_variant_set_fields(
index: RegistrarIndex,
fields: IdentityFields
) -> Self

Create a call with the variant set_fields.

pub fn new_call_variant_provide_judgement(
reg_index: RegistrarIndex,
target: <<T as Config>::Lookup as StaticLookup>::Source,
judgement: Judgement<<<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>,
identity: T::Hash
) -> Self

Create a call with the variant provide_judgement.

pub fn new_call_variant_kill_identity(
target: <<T as Config>::Lookup as StaticLookup>::Source
) -> Self

Create a call with the variant kill_identity.

pub fn new_call_variant_add_sub(
sub: <<T as Config>::Lookup as StaticLookup>::Source,
data: Data
) -> Self

Create a call with the variant add_sub.

pub fn new_call_variant_rename_sub(
sub: <<T as Config>::Lookup as StaticLookup>::Source,
data: Data
) -> Self

Create a call with the variant rename_sub.

pub fn new_call_variant_remove_sub(
sub: <<T as Config>::Lookup as StaticLookup>::Source
) -> Self

Create a call with the variant remove_sub.

pub fn new_call_variant_quit_sub() -> Self

Create a call with the variant quit_sub.