pallet_identity::pallet

Struct Pallet

pub struct Pallet<T>(/* private fields */);

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Implementations

impl<T: Config> Pallet<T>

Identity pallet declaration.

pub fn add_registrar( origin: OriginFor<T>, account: <<T as Config>::Lookup as StaticLookup>::Source, ) -> DispatchResultWithPostInfo

Add a registrar to the system.

The dispatch origin for this call must be T::RegistrarOrigin.

account: the account of the registrar.

Emits RegistrarAdded if successful.

pub fn set_identity( origin: OriginFor<T>, info: Box<T::IdentityInformation>, ) -> DispatchResultWithPostInfo

Set an account’s identity information and reserve the appropriate deposit.

If the account already has identity information, the deposit is taken as part payment for the new deposit.

The dispatch origin for this call must be Signed.

info: The identity information.

Emits IdentitySet if successful.

pub fn set_subs( origin: OriginFor<T>, subs: Vec<(T::AccountId, Data)>, ) -> DispatchResultWithPostInfo

Set the sub-accounts of the sender.

Payment: Any aggregate balance reserved by previous set_subs calls will be returned and an amount SubAccountDeposit will be reserved for each item in subs.

The dispatch origin for this call must be Signed and the sender must have a registered identity.

subs: The identity’s (new) sub-accounts.

pub fn clear_identity(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Clear an account’s identity info and all sub-accounts and return all deposits.

Payment: All reserved balances on the account are returned.

The dispatch origin for this call must be Signed and the sender must have a registered identity.

Emits IdentityCleared if successful.

pub fn request_judgement( origin: OriginFor<T>, reg_index: RegistrarIndex, max_fee: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, ) -> DispatchResultWithPostInfo

Request a judgement from a registrar.

Payment: At most max_fee will be reserved for payment to the registrar if judgement given.

The dispatch origin for this call must be Signed and the sender must have a registered identity.

reg_index: The index of the registrar whose judgement is requested.

max_fee: The maximum fee that may be paid. This should just be auto-populated as:

Registrars::<T>::get().get(reg_index).unwrap().fee

Emits JudgementRequested if successful.

pub fn cancel_request( origin: OriginFor<T>, reg_index: RegistrarIndex, ) -> DispatchResultWithPostInfo

Cancel a previous request.

Payment: A previously reserved deposit is returned on success.

The dispatch origin for this call must be Signed and the sender must have a registered identity.

reg_index: The index of the registrar whose judgement is no longer requested.

Emits JudgementUnrequested if successful.

pub fn set_fee( origin: OriginFor<T>, index: RegistrarIndex, fee: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, ) -> DispatchResultWithPostInfo

Set the fee required for a judgement to be requested from a registrar.

The dispatch origin for this call must be Signed and the sender must be the account of the registrar whose index is index.

index: the index of the registrar whose fee is to be set.

fee: the new fee.

pub fn set_account_id( origin: OriginFor<T>, index: RegistrarIndex, new: <<T as Config>::Lookup as StaticLookup>::Source, ) -> DispatchResultWithPostInfo

Change the account associated with a registrar.

The dispatch origin for this call must be Signed and the sender must be the account of the registrar whose index is index.

index: the index of the registrar whose fee is to be set.

new: the new account ID.

pub fn set_fields( origin: OriginFor<T>, index: RegistrarIndex, fields: <T::IdentityInformation as IdentityInformationProvider>::FieldsIdentifier, ) -> DispatchResultWithPostInfo

Set the field information for a registrar.

The dispatch origin for this call must be Signed and the sender must be the account of the registrar whose index is index.

index: the index of the registrar whose fee is to be set.

fields: the fields that the registrar concerns themselves with.

pub fn provide_judgement( origin: OriginFor<T>, reg_index: RegistrarIndex, target: <<T as Config>::Lookup as StaticLookup>::Source, judgement: Judgement<<<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>, identity: T::Hash, ) -> DispatchResultWithPostInfo

Provide a judgement for an account’s identity.

The dispatch origin for this call must be Signed and the sender must be the account of the registrar whose index is reg_index.

reg_index: the index of the registrar whose judgement is being made.

target: the account whose identity the judgement is upon. This must be an account with a registered identity.

judgement: the judgement of the registrar of index reg_index about target.

identity: The hash of the IdentityInformationProvider for that the judgement is provided.

Note: Judgements do not apply to a username.

Emits JudgementGiven if successful.

pub fn kill_identity( origin: OriginFor<T>, target: <<T as Config>::Lookup as StaticLookup>::Source, ) -> DispatchResultWithPostInfo

Remove an account’s identity and sub-account information and slash the deposits.

Payment: Reserved balances from set_subs and set_identity are slashed and handled by Slash. Verification request deposits are not returned; they should be cancelled manually using cancel_request.

The dispatch origin for this call must match T::ForceOrigin.

target: the account whose identity the judgement is upon. This must be an account with a registered identity.

Emits IdentityKilled if successful.

pub fn add_sub( origin: OriginFor<T>, sub: <<T as Config>::Lookup as StaticLookup>::Source, data: Data, ) -> DispatchResult

Add the given account to the sender’s subs.

Payment: Balance reserved by a previous set_subs call for one sub will be repatriated to the sender.

The dispatch origin for this call must be Signed and the sender must have a registered sub identity of sub.

pub fn rename_sub( origin: OriginFor<T>, sub: <<T as Config>::Lookup as StaticLookup>::Source, data: Data, ) -> DispatchResult

Alter the associated name of the given sub-account.

The dispatch origin for this call must be Signed and the sender must have a registered sub identity of sub.

pub fn remove_sub( origin: OriginFor<T>, sub: <<T as Config>::Lookup as StaticLookup>::Source, ) -> DispatchResult

Remove the given account from the sender’s subs.

Payment: Balance reserved by a previous set_subs call for one sub will be repatriated to the sender.

The dispatch origin for this call must be Signed and the sender must have a registered sub identity of sub.

pub fn quit_sub(origin: OriginFor<T>) -> DispatchResult

Remove the sender as a sub-account.

Payment: Balance reserved by a previous set_subs call for one sub will be repatriated to the sender (not the original depositor).

The dispatch origin for this call must be Signed and the sender must have a registered super-identity.

NOTE: This should not normally be used, but is provided in the case that the non- controller of an account is maliciously registered as a sub-account.

pub fn add_username_authority( origin: OriginFor<T>, authority: <<T as Config>::Lookup as StaticLookup>::Source, suffix: Vec<u8>, allocation: u32, ) -> DispatchResult

Add an AccountId with permission to grant usernames with a given suffix appended.

The authority can grant up to allocation usernames. To top up the allocation or change the account used to grant usernames, this call can be used with the updated parameters to overwrite the existing configuration.

pub fn remove_username_authority( origin: OriginFor<T>, suffix: Vec<u8>, authority: <<T as Config>::Lookup as StaticLookup>::Source, ) -> DispatchResult

Remove authority from the username authorities.

pub fn set_username_for( origin: OriginFor<T>, who: <<T as Config>::Lookup as StaticLookup>::Source, username: Vec<u8>, signature: Option<T::OffchainSignature>, use_allocation: bool, ) -> DispatchResult

Set the username for who. Must be called by a username authority.

If use_allocation is set, the authority must have a username allocation available to spend. Otherwise, the authority will need to put up a deposit for registering the username.

Users can either pre-sign their usernames or accept them later.

Usernames must:

Only contain lowercase ASCII characters or digits.

When combined with the suffix of the issuing authority be less than the MaxUsernameLength.

pub fn accept_username( origin: OriginFor<T>, username: BoundedVec<u8, <T as Config>::MaxUsernameLength>, ) -> DispatchResultWithPostInfo

Accept a given username that an authority granted. The call must include the full username, as in username.suffix.

pub fn remove_expired_approval( origin: OriginFor<T>, username: BoundedVec<u8, <T as Config>::MaxUsernameLength>, ) -> DispatchResultWithPostInfo

Remove an expired username approval. The username was approved by an authority but never accepted by the user and must now be beyond its expiration. The call must include the full username, as in username.suffix.

pub fn set_primary_username( origin: OriginFor<T>, username: BoundedVec<u8, <T as Config>::MaxUsernameLength>, ) -> DispatchResult

Set a given username as the primary. The username should include the suffix.

pub fn unbind_username( origin: OriginFor<T>, username: BoundedVec<u8, <T as Config>::MaxUsernameLength>, ) -> DispatchResult

Start the process of removing a username by placing it in the unbinding usernames map. Once the grace period has passed, the username can be deleted by calling remove_username.

pub fn remove_username( origin: OriginFor<T>, username: BoundedVec<u8, <T as Config>::MaxUsernameLength>, ) -> DispatchResultWithPostInfo

Permanently delete a username which has been unbinding for longer than the grace period. Caller is refunded the fee if the username expired and the removal was successful.

pub fn kill_username( origin: OriginFor<T>, username: BoundedVec<u8, <T as Config>::MaxUsernameLength>, ) -> DispatchResultWithPostInfo

Call with ForceOrigin privileges which deletes a username and slashes any deposit associated with it.

impl<T: Config> Pallet<T>

pub fn subs(who: &T::AccountId) -> Vec<(T::AccountId, Data)>

Get the subs of an account.

pub fn has_identity( who: &T::AccountId, fields: <T::IdentityInformation as IdentityInformationProvider>::FieldsIdentifier, ) -> bool

Check if the account has corresponding identity information by the identity field.

pub fn validate_signature( data: &[u8], signature: &T::OffchainSignature, signer: &T::AccountId, ) -> DispatchResult

Validate a signature. Supports signatures on raw data or data wrapped in HTML <Bytes>.

pub fn insert_username( who: &T::AccountId, username: BoundedVec<u8, <T as Config>::MaxUsernameLength>, provider: Provider<<<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>, )

A username has met all conditions. Insert the relevant storage items.

pub fn queue_acceptance( who: &T::AccountId, username: BoundedVec<u8, <T as Config>::MaxUsernameLength>, provider: Provider<<<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>, )

A username was granted by an authority, but must be accepted by who. Put the username into a queue for acceptance.

pub fn reap_identity( who: &T::AccountId, ) -> Result<(u32, u32, u32), DispatchError>

Reap an identity, clearing associated storage items and refunding any deposits. This function is very similar to (a) clear_identity, but called on a target account instead of self; and (b) kill_identity, but without imposing a slash.

Parameters:

target: The account for which to reap identity state.

Return type is a tuple of the number of registrars, IdentityInfo bytes, and sub accounts, respectively.

NOTE: This function is here temporarily for migration of Identity info from the Polkadot Relay Chain into a system parachain. It will be removed after the migration.

pub fn poke_deposit( target: &T::AccountId, ) -> Result<(<<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance), DispatchError>

Update the deposits held by target for its identity info.

Parameters:

target: The account for which to update deposits.

Return type is a tuple of the new Identity and Subs deposits, respectively.

NOTE: This function is here temporarily for migration of Identity info from the Polkadot Relay Chain into a system parachain. It will be removed after the migration.

pub fn set_identity_no_deposit( who: &T::AccountId, info: T::IdentityInformation, ) -> DispatchResult

Set an identity with zero deposit. Used for benchmarking and XCM emulator tests that involve rejig_deposit.

pub fn set_subs_no_deposit( who: &T::AccountId, subs: Vec<(T::AccountId, Data)>, ) -> DispatchResult

Set subs with zero deposit and default name. Only used for benchmarks that involve rejig_deposit.