Type Alias pallet_identity::pallet::Registrars

pub type Registrars<T: Config> = StorageValue<_GeneratedPrefixForStorageRegistrars<T>, BoundedVec<Option<RegistrarInfo<<<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, T::AccountId, <T::IdentityInformation as IdentityInformationProvider>::FieldsIdentifier>>, T::MaxRegistrars>, ValueQuery>;

The set of registrars. Not expected to get very big as can only be added through a special origin (likely a council motion).

The index into this can be cast to RegistrarIndex to get a valid value.

Storage type is StorageValue with value type `BoundedVec < Option < RegistrarInfo < BalanceOf < T > , T :: AccountId, < T :: IdentityInformation as IdentityInformationProvider > :: FieldsIdentifier, > ,

, T :: MaxRegistrars, >`.

Aliased Type

struct Registrars<T: Config>(/* private fields */);