Trait pallet_identity::Trait

pub trait Trait: Trait {
    type Event: From<Event<Self>> + Into<Self::Event>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type BasicDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type FieldDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type SubAccountDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type MaximumSubAccounts: Get<u32>;
    type Slashed: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;
    type ForceOrigin: EnsureOrigin<Self::Origin>;
    type RegistrarOrigin: EnsureOrigin<Self::Origin>;
}

Associated Types

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

type Currency: ReservableCurrency<Self::AccountId>

The currency trait.

type BasicDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount held on deposit for a registered identity.

type FieldDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount held on deposit per additional field for a registered identity.

type SubAccountDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount held on deposit for a registered subaccount. This should account for the fact that one storage item's value will increase by the size of an account ID, and there will be another trie item whose value is the size of an account ID plus 32 bytes.

type MaximumSubAccounts: Get<u32>

The maximum number of sub-accounts allowed per identified account.

type Slashed: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

What to do with slashed funds.

type ForceOrigin: EnsureOrigin<Self::Origin>

The origin which may forcibly set or remove a name. Root can always do this.

type RegistrarOrigin: EnsureOrigin<Self::Origin>

The origin which may add or remove registrars. Root can always do this.

Loading content...