Module pallet_identity::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

accept_username

Accept a given username that an authority granted. The call must include the full username, as in username.suffix.

add_registrar

Add a registrar to the system.

add_sub

Add the given account to the sender’s subs.

add_username_authority

Add an AccountId with permission to grant usernames with a given suffix appended.

cancel_request

Cancel a previous request.

clear_identity

Clear an account’s identity info and all sub-accounts and return all deposits.

kill_identity

Remove an account’s identity and sub-account information and slash the deposits.

kill_username

Call with ForceOrigin privileges which deletes a username and slashes any deposit associated with it.

provide_judgement

Provide a judgement for an account’s identity.

quit_sub

Remove the sender as a sub-account.

remove_expired_approval

Remove an expired username approval. The username was approved by an authority but never accepted by the user and must now be beyond its expiration. The call must include the full username, as in username.suffix.

remove_sub

Remove the given account from the sender’s subs.

remove_username

Permanently delete a username which has been unbinding for longer than the grace period. Caller is refunded the fee if the username expired and the removal was successful.

remove_username_authority

Remove authority from the username authorities.

rename_sub

Alter the associated name of the given sub-account.

request_judgement

Request a judgement from a registrar.

set_account_id

Change the account associated with a registrar.

set_fee

Set the fee required for a judgement to be requested from a registrar.

set_fields

Set the field information for a registrar.

set_identity

Set an account’s identity information and reserve the appropriate deposit.

set_primary_username

Set a given username as the primary. The username should include the suffix.

set_subs

Set the sub-accounts of the sender.

set_username_for

Set the username for who. Must be called by a username authority.

unbind_username

Start the process of removing a username by placing it in the unbinding usernames map. Once the grace period has passed, the username can be deleted by calling remove_username.