Module pallet_identity::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

add_registrar

Add a registrar to the system.

add_sub

Add the given account to the sender’s subs.

cancel_request

Cancel a previous request.

clear_identity

Clear an account’s identity info and all sub-accounts and return all deposits.

kill_identity

Remove an account’s identity and sub-account information and slash the deposits.

provide_judgement

Provide a judgement for an account’s identity.

quit_sub

Remove the sender as a sub-account.

remove_sub

Remove the given account from the sender’s subs.

rename_sub

Alter the associated name of the given sub-account.

request_judgement

Request a judgement from a registrar.

set_account_id

Change the account associated with a registrar.

set_fee

Set the fee required for a judgement to be requested from a registrar.

set_fields

Set the field information for a registrar.

set_identity

Set an account’s identity information and reserve the appropriate deposit.

set_subs

Set the sub-accounts of the sender.