Module pallet_identity::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

AccountOfUsername

Reverse lookup from username to the AccountId that has registered it. The value should be a key in the IdentityOf map, but it may not if the user has cleared their identity.

IdentityOf

Information that is pertinent to identify the entity behind an account. First item is the registration, second is the account’s primary username.

PendingUsernames

Usernames that an authority has granted, but that the account controller has not confirmed that they want it. Used primarily in cases where the AccountId cannot provide a signature because they are a pure proxy, multisig, etc. In order to confirm it, they should call Call::accept_username.

Registrars

The set of registrars. Not expected to get very big as can only be added through a special origin (likely a council motion).

SubsOf

Alternative “sub” identities of this account.

SuperOf

The super-identity of an alternative “sub” identity together with its name, within that context. If the account is not some other account’s sub-identity, then just None.

UsernameAuthorities

A map of the accounts who are authorized to grant usernames.