Trait pallet_identity::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type BasicDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type FieldDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type SubAccountDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type MaxSubAccounts: Get<u32>;
    type MaxAdditionalFields: Get<u32>;
    type MaxRegistrars: Get<u32>;
    type Slashed: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance>;
    type ForceOrigin: EnsureOrigin<Self::Origin>;
    type RegistrarOrigin: EnsureOrigin<Self::Origin>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type Currency: ReservableCurrency<Self::AccountId>

The currency trait.

type BasicDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The amount held on deposit for a registered identity

type FieldDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The amount held on deposit per additional field for a registered identity.

type SubAccountDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The amount held on deposit for a registered subaccount. This should account for the fact that one storage item’s value will increase by the size of an account ID, and there will be another trie item whose value is the size of an account ID plus 32 bytes.

type MaxSubAccounts: Get<u32>

The maximum number of sub-accounts allowed per identified account.

type MaxAdditionalFields: Get<u32>

Maximum number of additional fields that may be stored in an ID. Needed to bound the I/O required to access an identity, but can be pretty high.

type MaxRegistrars: Get<u32>

Maxmimum number of registrars allowed in the system. Needed to bound the complexity of, e.g., updating judgements.

type Slashed: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance>

What to do with slashed funds.

type ForceOrigin: EnsureOrigin<Self::Origin>

The origin which may forcibly set or remove a name. Root can always do this.

type RegistrarOrigin: EnsureOrigin<Self::Origin>

The origin which may add or remove registrars. Root can always do this.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.