Trait pallet_identity::IdentityInformationProvider

pub trait IdentityInformationProvider:
    Encode
    + Decode
    + MaxEncodedLen
    + Clone
    + Debug
    + Eq
    + PartialEq
    + TypeInfo
    + Default {
    type FieldsIdentifier: Member + Encode + Decode + MaxEncodedLen + TypeInfo + Default;

    // Required method
    fn has_identity(&self, fields: Self::FieldsIdentifier) -> bool;
}

Information concerning the identity of the controller of an account.

Required Associated Types

type FieldsIdentifier: Member + Encode + Decode + MaxEncodedLen + TypeInfo + Default

Type capable of holding information on which identity fields are set.

Required Methods

fn has_identity(&self, fields: Self::FieldsIdentifier) -> bool

Check if an identity registered information for some given fields.

Object Safety

This trait is not object safe.