Trait pallet_identity::IdentityInformationProvider

pub trait IdentityInformationProvider: Encode + Decode + MaxEncodedLen + Clone + Debug + Eq + PartialEq + TypeInfo {
    type IdentityField: Clone + Debug + Eq + PartialEq + TypeInfo + U64BitFlag;

    // Required method
    fn has_identity(&self, fields: u64) -> bool;

    // Provided method
    fn additional(&self) -> usize { ... }
}

Information concerning the identity of the controller of an account.

Required Associated Types

type IdentityField: Clone + Debug + Eq + PartialEq + TypeInfo + U64BitFlag

Type capable of representing all of the fields present in the identity information as bit flags in u64 format.

Required Methods

fn has_identity(&self, fields: u64) -> bool

Check if an identity registered information for some given fields.

Provided Methods

fn additional(&self) -> usize

👎Deprecated

Interface for providing the number of additional fields this identity information provider holds, used to charge for additional storage and weight. This interface is present for backwards compatibility reasons only and will be removed as soon as the reference identity provider removes additional fields.